LANG-SURFACE-INT-PRECISION: the surface reaches the kernel's arbitrary-precision Int
`LANG-SURFACE-INT-PRECISION` — the surface reaches the kernel's arbitrary-precision `Int`. Decision `dec_2qq0wcwbmjkvy` resolved APPROVE; Architect `evt_6w6jbc8gcgb0x`, QA `evt_1stn174df8hem`.

One commit, 41 `crates/**` paths, `+324/-85`. Merge-base `3ae945bd`.

## What lands

`Int` is specified arbitrary-precision, `OQ-int` was DECIDED 2026-06-27, and the kernel already carried `Term::IntLit(num_bigint::BigInt)`. The surface truncated it to `i128` through a lossy `n as i128`. **That truncation is gone, and gone from every carrier** — including `NumericLitVal::Int` and `int_lit_val`, which sit on the *evaluation* exit rather than the kernel one and which the original frame missed.

**The two exits are both measured.** A literal beyond `i128::MAX` reaches the kernel as an `IntLit` with the value written, and evaluates to the written value. Those are different code paths; proving only the kernel side would have proved the half that was never going to be the problem.

## The representability guard

Added mid-flight by Architect ruling `evt_3zj3z2x1get7p`, atomically part of this node rather than deferred.

Widening the carrier **newly admits** a counterexample: `const x : UInt8 = 2^128` is refused today at `lexer.rs:575` as unrepresentable in `i128`, and after widening would have flowed to `(n as u8)` and silently become `0`. A node that widens an acceptance surface owns the values it newly accepts.

So a declaration-owned `(type id, min, max)` descriptor now decides `min <= n <= max` on the mathematical `BigInt` **before** the literal primitive is created. In range, the exact value; out of range, a span-bearing error naming the literal, the target type, and the representable interval. No overflow VC and no runtime check — a closed constant's representability is decidable at elaboration.

The boundary discriminators hold in both directions: `255`/`256 : UInt8`, `127`/`128 : Int8`, `18446744073709551615 : UInt64` and its successor, with causal production mutations rather than a test-only oracle.

## Excluded and absent

No numeric tower, no `Decimal` mantissa work (that is `LANG-SURFACE-DECIMAL-PRECISION`), no `Float`/`Float32` change, no `OQ-1a` overflow-obligation work, no performance work, no new surface production, no `spec/` edit, and `trusted_base()` unchanged.

+++ crates/ken-elaborator/src/ast.rs
@@ -2,6 +2,7 @@
 //! `34` — data/match/refinements).
 
 use crate::error::Span;
+use num_bigint::BigInt;
 
 /// A binder group `(x y z : A)` from a `view` parameter list.
 #[derive(Clone, Debug)]
@@ -535,7 +536,7 @@ impl Decl {
 #[derive(Clone, Debug, PartialEq)]
 pub enum NumLit {
     /// Integer literal — defaults to `Int` unless an expected type is given.
-    Int(i128),
+    Int(BigInt),
     /// Decimal-point float — defaults to `Float`.
     Float(f64),
     /// `d`-suffix exact decimal — defaults to `Decimal`.
